+++ CHANGES.md
@@ -35,6 +35,7 @@ Changes in 7.0
 - JS MarkdownToHTML No MarkDown in text, return raw text in warehouse.js
 - Fix Delete from other Student/User Info tabs in Student.php & User.php
 - Remove deprecated since 4.5 rollover_* action hooks in Rollover.php & Actions.php
+- Fix Error: There is no column for The value for 0. This value was not saved in SaveData.fnc.php
 
 Changes in 6.9.4
 ----------------